+++ src/commands/catalog.rs
