+++ src/commands/cheat.rs
