+++ src/commands/inspect.rs
