+++ src/main.rs
@@ -4,22 +4,14 @@
 #![allow(clippy::enum_glob_use)]
 
 mod args;
-mod build;
-mod catalog;
-mod cheat;
 mod cli;
+mod commands;
 mod config;
 mod crypto;
-mod export;
 mod file_names;
 mod images;
-mod import;
-mod inspect;
-mod keys;
 mod langs;
-mod monitor;
 mod net;
-mod repl;
 mod repl_helper;
 mod vfs;
 mod wasm;
